requirements.txt
streamlit
google-generativeai
# sqlite3
python-dotenv